move datatableproviders dir under ajax
